requirements.txt
nats-py>=2.4.0
watchdog>=3.0.0